Fix bulk edit form
