+++ app/helpers/simple_form_helper.rb
@@ -1,10 +1,10 @@
 # frozen_string_literal: true
 
-# Ensures all Simple Form forms use BootstrapSwitcherFormBuilder unless explicitly overridden.
+# Ensures all Simple Form forms use CustomSimpleForm::FormBuilder unless explicitly overridden.
 # This allows us to switch between Bootstrap 4 and 5 wrappers based on the Current.bs5 flag
 module SimpleFormHelper
   def simple_form_for(record, options = {}, &)
-    options[:builder] ||= BootstrapSwitcherFormBuilder
+    options[:builder] ||= CustomSimpleForm::FormBuilder
     super
   end
 end
